use native lib
