+++ index.js
@@ -1,16 +1,5 @@
-
-
-var Cipher = require('./aead');
-exports.aead = Cipher;
-exports.createCipher = createCipher;
-function createCipher(key, iv) {
-  return new Cipher(key, iv);
+try {
+	module.exports = require('chacha-native');
+} catch (_) {
+	module.exports = require('./browser');
 }
-exports.createDecipher = createDecipher;
-function createDecipher(key, iv) {
-  return new Cipher(key, iv, true);
-}
-
-exports.createHmac = require('./polystream');
-exports.chacha20 = exports.ChaCha20 = require('./chachastream');
-exports.aeadLegacy = exports.AeadLegacy = require('./legacy-aead');
